+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## ADDED
 
+- Creation of the Cell class
+- Creation of the add_letter and calculate_value methods
+
+## [0.0.4] - 23-08-2023
+
+## ADDED
+
 - Creation of the Player class
 
 ## FIXED